requirements.txt
fastapi==2.2.2
uvicorn==0.40.0
sqlmodel==0.0.31
dotenv==0.9.9
psycopg2==2.9.11
bcrypt==5.0.0
pyjwt==2.10.1
aioboto3==15.5.0
python-multipart==0.0.22
ollama==0.6.1
cerebras_cloud_sdk==1.67.0
pypandoc==1.16.2
python-docx==1.2.0
pgvector==0.4.2
langchain-text-splitters==1.1.1